# api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://127.0.0.1:8080
    description: Generated server url
paths:
...
  /api/resCompanies/{id}:
    get:
      tags:
        - res-company-resource
      operationId: getResCompany
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResCompanyDTO"
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    put:
      tags:
        - res-company-resource
      operationId: updateResCompany
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResCompanyDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    delete:
      tags:
        - res-company-resource
      operationId: deleteResCompany
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/hrJobTitles/{id}:
    get:
      tags:
        - hr-job-title-resource
      operationId: getHrJobTitle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HrJobTitleDTO"
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    put:
      tags:
        - hr-job-title-resource
      operationId: updateHrJobTitle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HrJobTitleDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    delete:
      tags:
        - hr-job-title-resource
      operationId: deleteHrJobTitle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/hrJobPositions/{id}:
    get:
      tags:
        - hr-job-position-resource
      operationId: getHrJobPosition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HrJobPositionDTO"
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    put:
      tags:
        - hr-job-position-resource
      operationId: updateHrJobPosition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HrJobPositionDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    delete:
      tags:
        - hr-job-position-resource
      operationId: deleteHrJobPosition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /api/resCompanies:
    get:
      tags:
        - res-company-resource
      operationId: getAllResCompanies
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResCompanyDTO"
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    post:
      tags:
        - res-company-resource
      operationId: createResCompany
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResCompanyDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/hrJobTitles:
    get:
      tags:
        - hr-job-title-resource
      operationId: getAllHrJobTitles
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HrJobTitleDTO"
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    post:
      tags:
        - hr-job-title-resource
      operationId: createHrJobTitle
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HrJobTitleDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/hrJobPositions:
    get:
      tags:
        - hr-job-position-resource
      operationId: getAllHrJobPositions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HrJobPositionDTO"
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    post:
      tags:
        - hr-job-position-resource
      operationId: createHrJobPosition
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HrJobPositionDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "4xx/5xx":
          description: Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
components:
  schemas:
    ApiErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: "#/components/schemas/ApiFieldError"
    ApiFieldError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        property:
          type: string
        rejectedValue:
          type: object
        path:
          type: string
...
    ResCompanyDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 255
          minLength: 0
        address:
          type: string
          maxLength: 255
          minLength: 0
        phone:
          type: string
          maxLength: 255
          minLength: 0
        email:
          type: string
          maxLength: 255
          minLength: 0
        website:
          type: string
          maxLength: 255
          minLength: 0
        taxCode:
          type: string
          maxLength: 255
          minLength: 0
    HrJobTitleDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 255
          minLength: 0
        code:
          type: string
          maxLength: 255
          minLength: 0
        description:
          type: string
          maxLength: 255
          minLength: 0
      required:
        - code
        - name
    HrJobPositionDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 255
          minLength: 0
        targetRecruitment:
          type: integer
          format: int32
        jobSummary:
          type: string
        company:
          type: integer
          format: int64
        department:
          type: integer
          format: int64
        location:
          type: integer
          format: int64
        isRecruiting:
          type: boolean
      required:
        - company
...