# src/main/java/com/oop7hkem10/hrm/model/HrJobTitleDTO.java
package com.oop7hkem10.hrm.model;

import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.Getter;
import lombok.Setter;


@Getter
@Setter
public class HrJobTitleDTO {

    private Long id;

    @NotNull
    @Size(max = 255)
    private String name;

    @NotNull
    @Size(max = 255)
    @HrJobTitleCodeUnique
    private String code;

    @Size(max = 255)
    private String description;

}